% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moves}
\alias{moves}
\title{Moves}
\format{A data frame with 652 rows and 6 variables:
\describe{
  \item{name}{name of the move}
  \item{type}{type of the move}
  \item{category}{whether the move is a physical, special, or status move}
  \item{pp}{base Power Points of the move}
  \item{power}{Base Power, if it is a damaging move}
  \item{accuracy}{accuracy of the move (in percentage)}
}}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/List_of_moves_by_availability_(Generation_VIII)}
}
\usage{
moves
}
\description{
Pokémon moves available in Pokémon Sword & Shield.
}
\keyword{datasets}
