% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm_moves}
\alias{tm_moves}
\title{TM Moves}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{tmno}{number of the TM}
  \item{name}{name of the move}
}}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/TM}
}
\usage{
tm_moves
}
\description{
Moves that are available via Technical Machines (TM).
}
\keyword{datasets}
