% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr_moves}
\alias{tr_moves}
\title{TR Moves}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{trno}{number of the TR}
  \item{name}{name of the move}
}}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/TR}
}
\usage{
tr_moves
}
\description{
Moves that are available via Technical Records (TR).
}
\keyword{datasets}
