% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutor_moves}
\alias{tutor_moves}
\title{Tutor Moves}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{name}{name of the move}
  \item{location}{in-game location of the tutor teaching the move}
}}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/Move_Tutor}
}
\usage{
tutor_moves
}
\description{
Moves that are available via Move Tutor.
}
\keyword{datasets}
